tweak to DISABLE_AUTOSAVE flag
The flag used to be in void ProgramPage().

This caused a hangup in PumpFlashWrites(), when presets were switched. Instead I have moved the flag into the WritePattern() and WritePreset() function (the things we want to avoid writing when in demo mode), with the added bonus of still being able to record samples.

The DISABLE_AUTOSAVE flag is only relevant for demoing Plinky at shows like Superbooth but allows for a complete recall into a preset state on power cycle.